requirements.txt
streamlit==1.42.2opencv-python==4.11.0.86numpy==2.1.1torch==2.6.0torchvision==0.21.0ultralytics==8.3.78pillow==11.1.0